SecurityError is thrown if executing on FireFox
SecurityError is thrown if executing on FireFox, this is because the
code is attempting to access the cssRules attribute, which by default
returns "null" if you are not allowed to access, but FireFox throws an
exception. This fix aims to try/catch this exception and assign the
expected "null" to the variable. This issue only happens if there is an
external CSS referred in your HTML Dom